Qmake pro and compiler
The QMAKE pro file was updated to cross-compile the program from Linux to
OSX 10.9.  Special considerations have to be made to ensure that the qmake
properly informs OSX to the locations of the gsl/boost libraries.  The compiler
toolchain in QT Creator was updated to GCC4.8.1 downloaded from Macports.
Setting up a custo compiler on the mac osx system required creating a new QT
mkspecs folder within the QT directory that had reference to gcc48.