solution solver
Created the framework and implemented the bisection method.
If I was interested in using other methods then that can easily be done
once I create a function class that returns derivatives.  This was done
using the boost library but it would probably have been better to go with
GSL library.