Poptea framework
Moved the statistical methods into their own subclass. This was really messy
but this allows me to continue to expand the methods without polluting the
poptea public interface.  I need to think about more how to decide exactly
which objects to keep in the deeper thermal::analysis::methods section and
which to pop up into the poptea framework level.