Added best fit to the pie analysis
The bestfit is used to ensure that there are no residuals in the pie
analysis.  This issue was coming up in that the kernal and unknown
parameters were being corrupts away from the best-fit. This is because
the pie analysis modifies both the kernal and unknown parameters.
Alternatively, the kernal and unknown parameter objects can be copied
and saved and then loaded back in and passed back up out of pie.