Updates to thermalSweep
The updates to theraml sweep include the "addBefore" and "addAfter" methods. There is
now the requirement to feed in the thermal sweep for these methods.  This gives me
the ability to see which sweep is used for these initial and post data sets. The
thermal mapping variable iter has been moved so that it is declared later in the class.